Correct Jenkis Qulaité Gate

+++ src/main/java/com/test/spring/Jenkins/configuration/Calculator.java
@@ -15,6 +15,6 @@ public class Calculator {
         return a/b;
     }
     public static float sous(int a, int b ){
-        return  a-b;
+        return  (float)a-b;
     }
 }
